Remove one confusing data field in report
Remove the "invertedLatency" data field in the report. invertedLatency supposed
to mean 1 - latency / max(latency), and the purpose is to create an index that
represents the latency, and it's the higher the better just like the bandwidth.
But the name "invertedLatency" is very confusing, and It's not possible to find
a name that is short and explains this index. So the "invertedLatency" is
removed.